requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.23.2
pandas==2.1.1
google-cloud-storage==2.12.0
google-cloud-pubsub==3.15.0
openai==0.28.0
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
tenacity==9.2.2
python-multipart==0.0.6
